+++ src/modules/identity/endpoints/logout.endpoint.ts
@@ -2,7 +2,7 @@ import { Endpoint, HttpPost, Module } from '../../../../lib';
 
 @Module('auth')
 @HttpPost('logout')
-export class LogoutApi extends Endpoint {
+export class LogoutEndpoint extends Endpoint {
   main() {
     this.request.session.userId = undefined;
     return { ok: true };